requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
numpy
python-multipart
python-docx
pdfminer.six
PyPDF2
spacy
scikit-learn
sentence-transformers
transformers
rapidfuzz
pydantic
tqdm
python-dotenv
requests
